remove commands wip
