+++ src/SFA.DAS.EAS.Portal.UnitTests/Worker/EventHandlers/EventHandlerTestsFixture.cs
@@ -6,14 +6,13 @@ using System.Threading.Tasks;
 using AutoFixture;
 using Moq;
 using NServiceBus;
-using SFA.DAS.EAS.Portal.Application.Commands;
 using SFA.DAS.EAS.Portal.Application.Services;
 
 namespace SFA.DAS.EAS.Portal.UnitTests.Worker.EventHandlers
 {
-    public class EventHandlerTestsFixture<TEvent, TEventHandler, TCommand>
+    public class EventHandlerTestsFixture<TEvent, TEventHandler>
         where TEventHandler : IHandleMessages<TEvent>
-        where TCommand : class, ICommand<TEvent>
+        //where TCommand : class, ICommand<TEvent>
     {
         public TEvent Message { get; set; }
         public TEvent ExpectedMessage { get; set; }
@@ -21,9 +20,10 @@ namespace SFA.DAS.EAS.Portal.UnitTests.Worker.EventHandlers
         public string MessageId { get; set; }
         public Mock<IMessageContextInitialisation> MessageContextInitialisation { get; set; }
         public Mock<IMessageHandlerContext> MessageHandlerContext { get; set; }
-        public Mock<TCommand> Command { get; set; }
+        public Mock<IAccountDocumentService> AccountDocumentService { get; set; }
         
-        public EventHandlerTestsFixture(Func<IMessageContextInitialisation, IHandleMessages<TEvent>> constructHandler = null)
+//        public EventHandlerTestsFixture(Func<IAccountDocumentService, IMessageContextInitialisation, IHandleMessages<TEvent>> constructHandler = null)
+        public EventHandlerTestsFixture(bool constructHandler = true)
         {
             var fixture = new Fixture();
             Message = fixture.Create<TEvent>();
@@ -39,9 +39,11 @@ namespace SFA.DAS.EAS.Portal.UnitTests.Worker.EventHandlers
             messageHeaders.SetupGet(c => c["NServiceBus.TimeSent"]).Returns(DateTime.Now.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss:ffffff Z", CultureInfo.InvariantCulture));
             MessageHandlerContext.Setup(c => c.MessageHeaders).Returns(messageHeaders.Object);
 
-            Command = new Mock<TCommand>();
+            //Command = new Mock<TCommand>();
+            AccountDocumentService = new Mock<IAccountDocumentService>();
             
-            Handler = constructHandler != null ? constructHandler(MessageContextInitialisation.Object) : ConstructHandler();
+            if (constructHandler)
+                Handler = ConstructHandler();
         }
 
         public virtual Task Handle()
@@ -52,20 +54,20 @@ namespace SFA.DAS.EAS.Portal.UnitTests.Worker.EventHandlers
 
         private TEventHandler ConstructHandler()
         {
-            return (TEventHandler)Activator.CreateInstance(typeof(TEventHandler), Command.Object, MessageContextInitialisation.Object);
+            return (TEventHandler)Activator.CreateInstance(typeof(TEventHandler), AccountDocumentService.Object, MessageContextInitialisation.Object);
         }
 
-        public EventHandlerTestsFixture<TEvent, TEventHandler, TCommand> VerifyCommandExecutedWithUnchangedEvent()
-        {
-            Command.Verify(c => c.Execute(
-                It.Is<TEvent>(p => p.IsEqual(ExpectedMessage)), It.IsAny<CancellationToken>()),
-                Times.Once);
+//        public EventHandlerTestsFixture<TEvent, TEventHandler, TCommand> VerifyCommandExecutedWithUnchangedEvent()
+//        {
+//            Command.Verify(c => c.Execute(
+//                It.Is<TEvent>(p => p.IsEqual(ExpectedMessage)), It.IsAny<CancellationToken>()),
+//                Times.Once);
+//
+//            //todo: want test to fluent chain using derived methods, covariance?
+//            return this;
+//        }
 
-            //todo: want test to fluent chain using derived methods, covariance?
-            return this;
-        }
-
-        public EventHandlerTestsFixture<TEvent, TEventHandler, TCommand> VerifyMessageContextIsInitialised()
+        public EventHandlerTestsFixture<TEvent, TEventHandler> VerifyMessageContextIsInitialised()
         {
             MessageContextInitialisation.Verify(mc => mc.Initialise(MessageHandlerContext.Object), Times.Once);
