+++ src/SFA.DAS.EAS.Portal.UnitTests/Worker/EventHandlers/Reservations/ReservationCreatedEventHandlerTests.cs
@@ -1,6 +1,5 @@
 using System.Threading.Tasks;
 using NUnit.Framework;
-using SFA.DAS.EAS.Portal.Application.Commands;
 using SFA.DAS.EAS.Portal.Worker.EventHandlers.Reservations;
 using SFA.DAS.Reservations.Messages;
 using SFA.DAS.Testing;
@@ -17,15 +16,15 @@ namespace SFA.DAS.EAS.Portal.UnitTests.Worker.EventHandlers.Reservations
             return TestAsync(f => f.Handle(), f => f.VerifyMessageContextIsInitialised());
         }
         
-        [Test]
-        public Task Handle_WhenHandlingReservationCreatedEvent_ThenShouldExecuteAddReservationCommand()
-        {
-            return TestAsync(f => f.Handle(), f => f.VerifyCommandExecutedWithUnchangedEvent());
-        }
+//        [Test]
+//        public Task Handle_WhenHandlingReservationCreatedEvent_ThenShouldExecuteAddReservationCommand()
+//        {
+//            return TestAsync(f => f.Handle(), f => f.VerifyCommandExecutedWithUnchangedEvent());
+//        }
     }
 
     public class ReservationCreatedEventHandlerTestsFixture : EventHandlerTestsFixture<
-        ReservationCreatedEvent, ReservationCreatedEventHandler, ICommand<ReservationCreatedEvent>>
+        ReservationCreatedEvent, ReservationCreatedEventHandler>
     {
     }
 }
