+++ src/SFA.DAS.EAS.Portal.Worker/SFA.DAS.EAS.Portal.Worker.csproj
@@ -23,8 +23,10 @@
     <PackageReference Include="SFA.DAS.NLog.Targets.Redis" Version="1.2.1" />
     <PackageReference Include="SFA.DAS.NServiceBus" Version="10.0.1" />
     <PackageReference Include="SFA.DAS.ProviderRelationships.Messages" Version="3.0.2" />
+    <PackageReference Include="SFA.DAS.Providers.Api.Client" Version="0.11.174" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SFA.DAS.EAS.Portal.Client\SFA.DAS.EAS.Portal.Client.csproj" />
     <ProjectReference Include="..\SFA.DAS.EAS.Portal\SFA.DAS.EAS.Portal.csproj" />
   </ItemGroup>
   <ItemGroup>
