+++ src/SFA.DAS.EAS.Portal/Application/Commands/ICommand.cs
@@ -1,10 +1,10 @@
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.EAS.Portal.Application.Commands
-{
-    public interface ICommand<TParam>
-    {
-        Task Execute(TParam param, CancellationToken cancellationToken = default);
-    }
-}
+//using System.Threading;
+//using System.Threading.Tasks;
+//
+//namespace SFA.DAS.EAS.Portal.Application.Commands
+//{
+//    public interface ICommand<TParam>
+//    {
+//        Task Execute(TParam param, CancellationToken cancellationToken = default);
+//    }
+//}
