+++ src/SFA.DAS.EAS.Portal/DependencyResolution/ApplicationServices.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using SFA.DAS.EAS.Portal.Application.Commands;
 using SFA.DAS.EAS.Portal.Application.Services;
 
 namespace SFA.DAS.EAS.Portal.DependencyResolution
@@ -10,16 +9,17 @@ namespace SFA.DAS.EAS.Portal.DependencyResolution
         public static IServiceCollection AddApplicationServices(this IServiceCollection services, HostBuilderContext hostBuilderContext)
         {
             services.AddScoped<IMessageContext, MessageContext>();            
+            services.AddScoped<IMessageContextInitialisation, MessageContext>(sp => (MessageContext)sp.GetService<IMessageContext>());
             services.AddTransient<IAccountDocumentService, AccountDocumentService>();                        
             services.Decorate<IAccountDocumentService, AccountDocumentServiceWithSetProperties>();
             services.Decorate<IAccountDocumentService, AccountDocumentServiceWithDuplicateCheck>();
 
-            services.Scan(scan =>
-                    scan.FromAssembliesOf(typeof(ICommand<>))
-                    .AddClasses(classes =>
-                    classes.AssignableTo(typeof(ICommand<>)).Where(_ => !_.IsGenericType))
-                        .AsImplementedInterfaces()
-                        .WithTransientLifetime());
+//            services.Scan(scan =>
+//                    scan.FromAssembliesOf(typeof(ICommand<>))
+//                    .AddClasses(classes =>
+//                    classes.AssignableTo(typeof(ICommand<>)).Where(_ => !_.IsGenericType))
+//                        .AsImplementedInterfaces()
+//                        .WithTransientLifetime());
 
             return services;
         }
